STEAM-3733 rename actions folder to requests
