+++ src/commands/api/get.js
@@ -1,7 +1,7 @@
 const { flags } = require('@oclif/command')
 const { getApiIdentifierArg, reqType, resolvedParam } = require('../../support/command/parse-input')
 const { pipeAsync } = require('../../utils/general')
-const { getApi } = require('../../actions/api')
+const { getApi } = require('../../requests/api')
 const { getResponseContent } = require('../../support/command/response-handler')
 const BaseCommand = require('../../support/command/base-command')
